Add files via upload
This is version 2 of the website. Contains all of the same information from the previous version and all of the required information request by Fabio. This version also includes a template from W3. A skeleton of the template can been seen in zzzskeleton.html which is not accessible by the site itself. Is included here in case anyone else wants to develop a separate page. styles.css is a css page designed by W3. Please, do NOT change anything in this file unless you want to break the website!